spark-hub-backend/app/crawler/requirements.txt
# 基础依赖
aiohttp==3.9.3
aiodns==3.1.1
uvloop==0.19.0
cchardet==2.1.7
python-dotenv==1.0.1
loguru==0.7.2

# 数据处理
lxml==5.1.0
beautifulsoup4==4.12.3
parsel==1.9.0

# 存储
motor==3.3.2
redis==5.0.1
boto3==1.34.49

# 监控
prometheus-client==0.20.0
psutil==5.9.8

# 代理支持
aiohttp-proxy==0.1.2
fake-useragent==1.4.0
aiohttp-socks==0.8.4
python-socks==2.4.4

# 动态渲染
playwright==1.42.0

# 反爬虫
user-agents==2.2.0
browser-cookie3==0.19.1
selenium-stealth==1.0.6
undetected-chromedriver==3.5.4
cloudscraper==1.2.71
pyppeteer-stealth==2.7.4

# 性能优化
ujson==5.9.0
orjson==3.9.15
msgpack==1.0.7
zstandard==0.22.0

# 开发工具
black==24.2.0
isort==5.13.2
mypy==1.8.0
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# 其他依赖
pymongo==4.6.1
aioredis==2.0.1
aiomysql==0.2.0
SQLAlchemy==2.0.25
pydantic==2.5.3
tenacity==8.2.3
aiofiles==23.2.1
aiokafka==0.10.0
mmh3==4.1.0
scrapy==2.11.0 